+++ Data_Collection/vdd.py
@@ -49,7 +49,7 @@ if __name__ == '__main__':
     RESTING = 0
 
     # Define accelerometers (named after rivers)
-    INDUS = raspi_accel_lib.ADXL345(0x1D)
+    INDUS = raspi_accel_lib.ADXL345(st.CAL_X, st.CAL_Y, st.CAL_Z, 0x1D)
 
     # Startup Accelerometer
     INDUS.accel_startup(st.GFORCE)